# swagger.yaml
openapi: 3.0.0
info:
  title: Email Validator API
  description: API to validate emails (syntax, MX, SMTP, SPF, DMARC, etc.)
  version: "1.1.7a"
  contact:
    email: contact.emailapi@gmail.com

paths:
  /check:
    get:
      summary: Validate a single email
      tags:
        - Email
      parameters:
        - name: email
          in: query
          required: true
          description: Email address to validate
          schema:
            type: string
            example: user@example.com
        - name: skipSMTP
          in: query
          required: false
          description: Skip SMTP probing (default false)
          schema:
            type: boolean
            example: true
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailCheckResult"
  /check-bulk:
    post:
      summary: Validate multiple emails in bulk
      tags:
        - Email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emails:
                  type: array
                  items:
                    type: string
                    example:
                      - user@example.com
                      - admin@mydomain.com
      responses:
        "200":
          description: Bulk validation result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailCheckResult"

components:
  schemas:
    EmailCheckResult:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        syntaxValid:
          type: boolean
          example: true
        mx:
          type: object
          nullable: true
          properties:
            ok:
              type: boolean
              example: true
        smtp:
          type: object
          nullable: true
          properties:
            ok:
              type: boolean
              example: false
            error:
              type: string
              nullable: true
              example: "Mailbox not found"
        inBlocklist:
          type: boolean
          example: false
        inTrustedDomains:
          type: boolean
          example: true
        isRole:
          type: boolean
          example: false
        isDisposable:
          type: boolean
          example: false
        hasSPF:
          type: boolean
          example: true
        hasDMARC:
          type: boolean
          example: true
        domainAgeYears:
          type: number
          example: 5
        noProbeList:
          type: boolean
          example: false
        requestsLeft:
          type: integer
          description: Remaining API requests (rate-limit info)
          example: 99